Fork the pom file and update the readme to account.
In particular, change the groupId of the fork, so when we absorb the upstream changes, we don't start erroneously attempting to push snapshots under com.squareup.  Additionally, if we have any changes that don't make it upstream, internal projects which depend on these changes will have a home.  Ideally we'll never release, and we'll only push snapshots to support 'code in transit' before it's uptreamed.