Drop name collision avoidance. Tweak ClassName API.
Name collision avoidance is difficult to do properly at this level
of abstraction. In particular, if we pick a different name for
something, it's awkward to surface that to the caller.

Also tweak the ClassName API. Use a single list for all simple
names, even the class' own simple name. This makes some of the
calling code simpler.