Small API ergonomics.
Don't require the caller to repeat the type when that's redundant.
addField(FieldSpec.of()) is now just add(FieldSpec.of()). This gives
us the option later of introducing a member type, and interleaving
members with a single add() method.

Also make it possible to write basic control flow without creating
an explicit code block.