Rename toBuilder() to builder().
I had a small insight: the only place where the static builder()
method conflicted with the non-static builder() method was in
CodeBlock. That class is composable: it's just as easy to create
a new code block, and call add(), as it is to convert an existing
code block into a builder. So we don't need the toBuilder() method
there anyway.