Add failing tests for types conflicting with type being declared
Also removes previous (flawed) test, which should have
been in JavaFileTest anyway, and add a separate test for
the "superclass references self" case (already handled
in TypeSpecTest#interfaceExtends btw, but this makes it
clearer what is tested and expected).
And add a separate test for classes in the same package
conflicting with a superclass, for completeness.

Reproduces #470