Add a `tap` method to builders
The builder pattern is nice, but it's frequent that you have some
conditional logic which forces you to interrupt the flow, assign
the builder to a temporary variable, do something on the builder,
then resume the fluent API like mode. A trivial example would be:

```java
TypeSpec.Builder typeBuilder = TypeSpec.classBuilder(...)
myMethods.forEach(typeBuilder::addMethod);
return typeBuilder.build()
```

This commit introduces a `tap` method on builders which avoids
the need of an intermediate variable by allowing to swich context
in a lambda, and apply the conditional logic (or more programmatic
work) within the lambda.

The code above would be rewritten as:

```java
return TypeSpec.classBuilder(...)
    .tap(c -> methods.forEach(c::addMethod))
    .build()
```

Which looks more functional style. Of course real world
examples the conditional logic is more complex ;)