Only build on Java 8
Running a build on different JDKs is pointless. Our users consume the artifact in binary form. We may want to run the tests on multiple JDKs. The way to do that is to migrate to Gradle, duplicate the test task, and use the toolchain feature. That is out of scope of this change, which is only to stop pointless CI builds.