chore: allow multithreaded use of $V
Previously, if multiple threads were inlining using the same
ObjectInliner, the internal state of the inliner would be corrupted.
This is fixed by creating a new ObjectEmitter whenever a value is
being inlined; the ObjectEmitters created from an ObjectInliner
are completely independent and thus cannot be corrupted by
multithreaded use. Also undos the unneccessary NameAllocator
change, update the API for TypeInliner, and adds more tests
for custom inlining.